+++ lib/face_liveness_detection_plus.dart
@@ -1,2 +1,3 @@
 export 'src/face_detection/face_detector.dart';
 export 'src/rule_set/rule_set.dart';
+export 'src/face_capture/face_capture_controller.dart';